Pass Python code to initialize from Java through C
Previously, we had some hard-coded Python code in the C part, but it is
much more flexible to store the Python code to start at spin-up time in
the Java resources.

For the moment, this code still only prints out a greeting; eventually,
it needs to verify that the code linking this JVM to Python is only run
when it was not executed already.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>