Use the build-helper to help Eclipse discover NarSystem
The NAR plugin generates the source code for the NarSystem class and puts
it somewhere else than Eclipse's Maven integration expects it. But there
is an m2e connector for the build-helper Maven plugin, so we can use --
just for Eclipse, my friend -- that plugin to make sure that that class is
found and no bogus compile error is reported.

When importing scripting-cpython afresh, m2e will ask whether you want to
discover connectors for Maven plugins (unless you already have the
build-helper one). Otherwise, you will have to hit Ctrl+1 on the line in
Eclipse's "Problems" view and install the connector then.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>